+++ packages/extension/src/background/index.js
@@ -6,3 +6,4 @@ export {
 export { createBackgroundHost } from './createBackgroundHost.js';
 export { attachChromeRuntime } from './ChromeRuntimeAdapter.js';
 export { attachSessionMetaListener } from './sessionMeta.js';
+export { ApprovalBroker } from './approvalBroker.js';